requirements.txt
aiohttp==3.8.4
discord.py==2.2.3
openai==0.27.7
pymongo==4.5.0
python-dotenv==1.0.0